feat: platform management
